requirements.txt
appdirs==1.4.4
ariadne==0.13.0
ariadne-django==0.2.0
asgiref==3.4.1
astroid==2.8.4
autoflake==1.4.0
black==21.5b2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
coverage==6.1.1
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-admin==2.0.1
django-allauth==0.45.0
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.12.4
empyrical==0.5.5
factory-boy==3.2.1
Faker==9.7.1
gunicorn==20.1.0
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pandas==1.3.4
pathspec==0.9.0
pre-commit==2.15.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.15.0
screen==1.0.1
sentry-sdk==1.4.3
six==1.16.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
xlwt==1.3.0
